+++ internal/handler/routes.go
@@ -29,6 +29,7 @@ func (s *Server) SetupRoutes(r *mux.Router) {
 	}
 	write.HandleFunc("/applications/stake", s.handleStakeNewApplication).Methods("POST")
 	write.HandleFunc("/applications/{address}/upstake", s.handleUpstake).Methods("POST")
+	write.HandleFunc("/applications/{address}/unstake", s.handleUnstake).Methods("POST")
 	write.HandleFunc("/applications/{address}/fund", s.handleFund).Methods("POST")
 	write.HandleFunc("/applications/{address}/delegate", s.handleDelegate).Methods("POST")
 	write.HandleFunc("/applications/{address}/autotopup", s.handleSetAutoTopUp).Methods("PUT")